+++ .gitignore
@@ -43,3 +43,4 @@ experiment/shaders/
 __pycache__/
 *__pycache__/
 *.pyc
+build/
