+++ CMakeLists.txt
@@ -4,10 +4,10 @@ project(openMTPK VERSION 0.1 LANGUAGES CXX)
 
 set(DEFAULT_BUILD_TYPE "Release")
 if(NOT CMAKE_BUILD_TYPE AND NOT CMAKE_CONFIGURATION_TYPES)
-  message(STATUS "Setting build type to '${DEFAULT_BUILD_TYPE}' as none was specified.")
-  set(CMAKE_BUILD_TYPE "${DEFAULT_BUILD_TYPE}" CACHE STRING "Choose the type of build." FORCE)
-  # Set the possible values of build type for cmake-gui
-  set_property(CACHE CMAKE_BUILD_TYPE PROPERTY STRINGS "Debug" "Release" "MinSizeRel" "RelWithDebInfo")
+    message(STATUS "Setting build type to '${DEFAULT_BUILD_TYPE}' as none was specified.")
+    set(CMAKE_BUILD_TYPE "${DEFAULT_BUILD_TYPE}" CACHE STRING "Choose the type of build." FORCE)
+    # Set the possible values of build type for cmake-gui
+    set_property(CACHE CMAKE_BUILD_TYPE PROPERTY STRINGS "Debug" "Release" "MinSizeRel" "RelWithDebInfo")
 endif()
 
 include(GNUInstallDirs)
@@ -22,29 +22,47 @@ set(SOURCE_FILES
     modules/ml/linreg.cpp
     modules/number_theory/cipher.cpp
     modules/number_theory/rc4.cpp
-    modules/number_theory/primes.cpp
 )
 
-add_library(${PROJECT_NAME} ...)
+add_library(${PROJECT_NAME} STATIC ${SOURCE_FILES})
 
-target_include_directories(${PROJECT_NAME} PUBLIC
+#target_sources(${PROJECT_NAME} PUBLIC FILE_SET HEADERS
+#    BASE_DIRS ${OPENMTPK_SOURCE_DIR}
+#    FILES ${OPENMTPK_SOURCE_DIR}/include
+#)
+
+target_include_directories(${PROJECT_NAME} PUBLIC 
     $<BUILD_INTERFACE:${CMAKE_CURRENT_SOURCE_DIR}/include>
     $<INSTALL_INTERFACE:include>
     PRIVATE modules
 )
 
-set_target_properties(${PROJECT_NAME} PROPERTIES
-    VERSION ${PROJECT_VERSION}
-    SOVERSION 1
+# headers to declare as public
+set(OPM_INCLUDES
+    ${PROJECT_SOURCE_DIR}/include/arithmetic/arith.hpp
+    ${PROJECT_SOURCE_DIR}/include/calculus/deriv.hpp
+    ${PROJECT_SOURCE_DIR}/include/linalg/lao.hpp
+    ${PROJECT_SOURCE_DIR}/include/ml/linreg.hpp
+    ${PROJECT_SOURCE_DIR}/include/number_theory/cipher.hpp
+    ${PROJECT_SOURCE_DIR}/include/number_theory/rc4.hpp
+    ${PROJECT_SOURCE_DIR}/include/number_theory/primes.hpp
 )
 
-install(TARGETS ${PROJECT_NAME} EXPORT openMTPKConfig
+set_target_properties(${PROJECT_NAME} 
+    PROPERTIES
+    PUBLIC_HEADER "${OPM_INCLUDES}"
+    VERSION ${PROJECT_VERSION}
+)
+
+install(TARGETS ${PROJECT_NAME} EXPORT openMTPKConfig 
     ARCHIVE  DESTINATION ${CMAKE_INSTALL_LIBDIR}
     LIBRARY  DESTINATION ${CMAKE_INSTALL_LIBDIR}
     RUNTIME  DESTINATION ${CMAKE_INSTALL_BINDIR}
+    PUBLIC_HEADER 
+        DESTINATION "${CMAKE_INSTALL_INCLUDEDIR}/openMTPK"
 )
 
-install(EXPORT opentMTPKConfig DESTINATION share/openMTPK/cmake)
+install(EXPORT openMTPKConfig DESTINATION share/openMTPK/cmake)
 
-export(TARGETS ${PROJECT_NAME} FILES opentMTPKConfig.cmake)
+export(TARGETS ${PROJECT_NAME} FILE openMTPKConfig.cmake)
