+++ all.sh
@@ -1,14 +1,4 @@
 #! /bin/sh
-#                        ___  ______________ _   __
-#                        |  \/  |_   _| ___ \ | / /
-#   ___  _ __   ___ _ __ | .  . | | | | |_/ / |/ / 
-#  / _ \| '_ \ / _ \ '_ \| |\/| | | | |  __/|    \ 
-# | (_) | |_) |  __/ | | | |  | | | | | |   | |\  \
-#  \___/| .__/ \___|_| |_\_|  |_/ \_/ \_|   \_| \_/
-#       | |                                        
-#       |_|                                        
-#
-#
 # this script acts as the main "runner" for openMTPK demonstrating its
 # functionality by running each modules driver file
