--- include/arithmetic/arithbase.hpp
@@ -1,61 +1,0 @@
-/*
- * definitions for basic arithmetic operations 
- */
-
-#ifndef ARITHBASE_HPP
-#define ARITHBASE_HPP
-#include <vector>
-#include <iostream>
-#include <stdio.h>
-#include <algorithm>
-
-class Arith {
-
-    public:
-        /*
-         * It is easiest to include the implementation of our template
-         * methods in our header files
-         * Addition
-         */
-        template<typename T>
-        inline T add(T t) {
-            return t;
-        }
-        template<typename T, typename... Ts>
-        inline auto add(T t, Ts... ts) {
-            return t + add(ts...);
-        }
-        
-        // subtraction
-        template<typename X>
-        inline X sub(X x) {
-            return x;
-        }
-        template<typename X, typename... Xy>
-        inline auto sub(X x, Xy... xy) {
-            return x - sub(xy...);
-        }
-
-        // multiplication
-        template<typename W>
-        inline W mult(W w) {
-            return w;
-        }
-        template<typename W, typename... Wv>
-        inline auto mult(W w, Wv... wv) {
-            return w * mult(wv...);
-        }
-
-        // exponentiation
-        template<typename Z>
-        inline Z exp(Z z) {
-            return z;
-        }
-        template<typename Z, typename... Zy>
-        inline auto exp(Z z, Zy... zy) {
-            return z *= exp(zy...);
-        }
-};
-
-#endif
-
