+++ modules/arithmetic/arith.cpp
@@ -1,7 +1,6 @@
 /*
  * stubbed out implementation file for arithmetic operations
  */
-
 #include "../../include/arithmetic/arith.hpp"
