+++ samples/cpp/cipher.cpp
@@ -6,6 +6,7 @@
 #include <string>
 #include <vector>
 #include "../../include/number_theory/cipher.hpp"
+//#include <openMTPK/cipher.hpp>
 
 
 int main() {
