--- tests/CMakeLists.txt
@@ -1,3 +1,0 @@
-
-
-
